ingest/ingest_college_football_data_api/requirements.lock.txt
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
idna==3.11
    # via requests
pyyaml==6.0.3
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
urllib3==2.6.3
    # via requests
